add collection feature and question explanation
合集(Collection)功能：
- collections/collection_courses/collection_enrollments 表 + 完整接口
- 教师端合集管理：新建/编辑/删除、添加课程、排序/移出、封面上传+AI生成
- 学生端：合集可选课进我的课程，合集详情整页展示课程
- 合集专属课程：从合集上传的新课程不出现在外部课程列表
- 合集分类标签

题目解析(explanation)端到端贯通：
- AI出题自动生成解析 + 手动出题解析输入栏
- 教师卡片/学生答题回顾显示解析

+++ app/src/main/java/com/studyapp/AdminActivity.kt
@@ -1467,7 +1467,7 @@ class AdminActivity : AppCompatActivity() {
         infoLayout.addView(subText)
 
         val masteryText = TextView(this).apply {
-            text = String.format("%.1f%%", mastery)
+            text = String.format("%.1f%%", mastery.toDouble())
             textSize = 14f
             setTypeface(null, Typeface.BOLD)
             setTextColor(resources.getColor(R.color.tsinghua_purple))
